Finished TicTacToe Project
Fixed win condition bug being set instead of checking (had = instead of ==).
Also added in the text "INPUT: " before the player types in the row and col. I think it helps to visualize where the player needs to type, and if they want to go back and read previous inputs it's a lot easier.