Completed TemplateFun project
Introduction to function templates and class templates.
getMax() and getMin() methods that return the max and min arguments respectively, regardless of data type (as long as it overrides the "<" or ">" operator).
Swapper class that holds two objects of the same data type and can swap the first object and second.